refactor: 공통 디자인 컴포넌트의 styled-components 파일 분리
[#806]
