+++ frontend/src/components/common/ButtonNext/ButtonNext.tsx
@@ -1,13 +1,10 @@
 import type { CSSProp } from 'styled-components';
-import styled, { css } from 'styled-components';
 
-import type { ButtonHTMLAttributes, ReactNode, MouseEventHandler } from 'react';
+import type { ButtonHTMLAttributes, MouseEventHandler, ReactNode } from 'react';
 import { forwardRef } from 'react';
 
+import { StyledButtonNext, StyledPureButton } from '@common/ButtonNext/ButtonNext.style';
 import type { SpacingProps } from '@common/systems';
-import { spacing } from '@common/systems';
-
-import { getColor, getHoverColor } from '@style';
 
 import type { Color, Size } from '@type/style';
 
@@ -26,101 +23,10 @@ export interface ButtonNextProps extends SpacingProps, ButtonHTMLAttributes<HTML
 
 const ButtonNext = ({ children, noTheme, ...props }: ButtonNextProps) => {
   return noTheme ? (
-    <S.PureButton {...props}>{children}</S.PureButton>
+    <StyledPureButton {...props}>{children}</StyledPureButton>
   ) : (
-    <S.Button {...props}>{children}</S.Button>
+    <StyledButtonNext {...props}>{children}</StyledButtonNext>
   );
-};
-
-const S = {
-  Button: styled.button<ButtonNextProps>`
-    border-radius: 6px;
-    ${({ pill }) => pill && 'border-radius: 20px;'}
-
-    ${({ fullWidth }) => fullWidth && 'width: 100%;'}
-    ${({ disabled }) => disabled && `cursor: unset;`}
-    ${({ variant, color, disabled }) => {
-      switch (variant) {
-        case 'text':
-          return css`
-            color: ${disabled ? '#a0a0a0' : color === 'light' ? '#000' : getColor(color)};
-            background: transparent;
-            border: none;
-
-            &:hover {
-              background: ${disabled ? 'transparent' : '#1976d20a'};
-            }
-          `;
-        case 'outlined':
-          return css`
-            color: ${disabled ? '#a0a0a0' : color === 'light' ? '#333' : getColor(color)};
-            background: transparent;
-            border: 1.5px solid ${disabled ? '#a0a0a0' : getColor(color)};
-
-            &:hover {
-              color: ${disabled ? '#a0a0a0' : color === 'light' ? '#333' : '#fff'};
-              background: ${disabled ? 'transparent' : getHoverColor(color)};
-            }
-          `;
-        case 'contained':
-        default:
-          return css`
-            color: ${disabled ? '#a0a0a0' : color === 'light' ? '#000' : '#ffffff'};
-            background: ${disabled ? '#e0e0e0' : getColor(color)};
-            border: 1.5px solid ${disabled ? '#e0e0e0' : getColor(color)};
-
-            &:hover {
-              background: ${disabled ? '#e0e0e0' : getHoverColor(color)};
-            }
-          `;
-      }
-    }}
-
-    padding: ${({ size }) => {
-      switch (size) {
-        case 'xs':
-          return '2px 8px';
-        case 'sm':
-          return '4px 12px';
-        case 'md':
-          return '6px 16px';
-        case 'lg':
-          return '8px 20px';
-        case 'xl':
-          return '10px 24px';
-        case 'xxl':
-          return '12px 28px';
-        default:
-          return '6px 16px';
-      }
-    }};
-
-    font-size: ${({ size }) => {
-      switch (size) {
-        case 'xs':
-          return '14px';
-        case 'sm':
-          return '16px';
-        case 'md':
-          return '18px';
-        case 'lg':
-          return '20px';
-        case 'xl':
-          return '22px';
-        case 'xxl':
-          return '24px';
-        default:
-          return '18px';
-      }
-    }};
-
-    ${spacing};
-
-    ${({ css }) => css};
-  `,
-  PureButton: styled.button<ButtonNextProps>`
-    ${({ css }) => css};
-  `,
 };
 
 export default forwardRef(ButtonNext);
