+++ frontend/src/components/common/FlexBox/FlexBox.tsx
@@ -1,21 +1,12 @@
 import type { CSSProp } from 'styled-components';
-import styled from 'styled-components';
 
 import { type HTMLAttributes, type ReactNode } from 'react';
 
+import type { FLEX_BOX_ITEM_POSITION } from '@common/FlexBox/FlexBox.style';
+import { StyledFlexBox } from '@common/FlexBox/FlexBox.style';
 import type { SpacingProps } from '@common/systems';
-import { spacing } from '@common/systems';
-
-import { borderRadius, getSize } from '@style';
 
 import type { AxisType, BorderRadiusDirectionType } from '@type/style';
-
-export const FLEX_BOX_ITEM_POSITION = {
-  start: 'start',
-  center: 'center',
-  end: 'end',
-  between: 'space-between',
-} as const;
 
 export interface FlexBoxProps extends HTMLAttributes<HTMLDivElement>, SpacingProps {
   tag?: string;
@@ -40,31 +31,6 @@ export interface FlexBoxProps extends HTMLAttributes<HTMLDivElement>, SpacingPro
   children: ReactNode;
 }
 
-export type StyledFlexBoxType = Omit<
-  FlexBoxProps,
-  | 'noRadius'
-  | 'rowGap'
-  | 'columnGap'
-  | 'justifyContent'
-  | 'alignItems'
-  | 'alignContent'
-  | 'minHeight'
-  | 'maxHeight'
-  | 'minWidth'
-  | 'maxWidth'
-> & {
-  $noRadius?: BorderRadiusDirectionType;
-  $rowGap?: number;
-  $columnGap?: number;
-  $justifyContent?: keyof typeof FLEX_BOX_ITEM_POSITION;
-  $alignItems?: keyof typeof FLEX_BOX_ITEM_POSITION;
-  $alignContent?: keyof typeof FLEX_BOX_ITEM_POSITION;
-  $minHeight?: number | string;
-  $maxHeight?: number | string;
-  $minWidth?: number | string;
-  $maxWidth?: number | string;
-};
-
 const FlexBox = ({
   children,
   tag,
@@ -83,7 +49,7 @@ const FlexBox = ({
   const changeableTag = tag || 'div';
 
   return (
-    <S.FlexBox
+    <StyledFlexBox
       as={changeableTag}
       $noRadius={noRadius}
       $rowGap={rowGap}
@@ -98,55 +64,8 @@ const FlexBox = ({
       {...props}
     >
       {children}
-    </S.FlexBox>
+    </StyledFlexBox>
   );
-};
-
-const getGap = ({ gap, rowGap, columnGap }: Pick<FlexBoxProps, 'gap' | 'rowGap' | 'columnGap'>) => {
-  if (gap !== undefined) {
-    return `${gap * 0.4}rem`;
-  }
-
-  const row = rowGap !== undefined ? rowGap * 0.4 : 0.4;
-  const column = columnGap !== undefined ? columnGap * 0.4 : 0.4;
-
-  return `${row}rem ${column}rem`;
-};
-
-const S = {
-  FlexBox: styled.div<StyledFlexBoxType>`
-    ${spacing};
-
-    width: ${({ width }) => getSize(width)};
-    min-width: ${({ $minWidth }) =>
-      typeof $minWidth === 'string' ? $minWidth : `${$minWidth}rem`};
-    max-width: ${({ $maxWidth }) =>
-      typeof $maxWidth === 'string' ? $maxWidth : `${$maxWidth}rem`};
-
-    height: ${({ height }) => getSize(height)};
-    min-height: ${({ $minHeight }) =>
-      typeof $minHeight === 'string' ? $minHeight : `${$minHeight}rem`};
-    max-height: ${({ $maxHeight }) =>
-      typeof $maxHeight === 'string' ? $maxHeight : `${$maxHeight}rem`};
-
-    flex-wrap: ${({ nowrap }) => (nowrap ? 'nowrap' : 'wrap')};
-    flex-direction: ${({ direction }) => (direction ? direction : 'row')};
-    justify-content: ${({ $justifyContent }) => FLEX_BOX_ITEM_POSITION[$justifyContent]};
-    align-items: ${({ $alignItems }) => FLEX_BOX_ITEM_POSITION[$alignItems]};
-    align-content: ${({ $alignContent }) => FLEX_BOX_ITEM_POSITION[$alignContent]};
-    gap: ${({ gap, $rowGap, $columnGap }) =>
-      getGap({ gap, rowGap: $rowGap, columnGap: $columnGap })};
-    ${({ background }) => background && `background: ${background};`}
-    border: ${({ outlined }) => (outlined ? '0.15rem solid #000' : 'none')};
-
-    display: flex;
-    border-radius: 1rem;
-    font-size: 1.5rem;
-
-    ${({ $noRadius }) => $noRadius && borderRadius($noRadius)};
-
-    ${({ css }) => css};
-  `,
 };
 
 export default FlexBox;
