+++ frontend/src/components/common/List/List.tsx
@@ -1,10 +1,9 @@
 import type { CSSProp } from 'styled-components';
-import styled from 'styled-components';
 
 import type { HTMLAttributes, ReactNode } from 'react';
 
+import { StyledList } from '@common/List/List.style';
 import type { SpacingProps } from '@common/systems';
-import { spacing } from '@common/systems';
 
 export interface ListProps extends HTMLAttributes<HTMLUListElement>, SpacingProps {
   children: ReactNode;
@@ -15,21 +14,10 @@ export interface ListProps extends HTMLAttributes<HTMLUListElement>, SpacingProp
 
 const List = ({ children, fontSize, ...props }: ListProps) => {
   return (
-    <ListWrapper $fontSize={fontSize} {...props}>
+    <StyledList $fontSize={fontSize} {...props}>
       {children}
-    </ListWrapper>
+    </StyledList>
   );
 };
-
-const ListWrapper = styled.ul<Omit<ListProps, 'fontSize'> & { $fontSize: number }>`
-  ${spacing}
-
-  list-style-type: none;
-  font-size: ${({ $fontSize }) => `${$fontSize}rem`};
-
-  ${({ border }) => border && `border: 0.01rem solid #66666666; border-radius:0.4rem;`}
-
-  ${({ css }) => css};
-`;
 
 export default List;
