+++ frontend/src/components/common/Loader/Loader.tsx
@@ -1,5 +1,6 @@
 import type { CSSProp } from 'styled-components';
-import styled from 'styled-components';
+
+import { StyledLoader } from '@common/Loader/Loader.style';
 
 import type { Size } from '@type';
 
@@ -9,60 +10,8 @@ export interface LoaderProps {
   css?: CSSProp;
 }
 
-const Loader = styled.div<LoaderProps>`
-  width: ${({ size }) => {
-    switch (size) {
-      case 'xs':
-        return '1.2rem';
-      case 'sm':
-        return '1.6rem';
-      case 'md':
-        return '2.0rem';
-      case 'lg':
-        return '2.4rem';
-      case 'xl':
-        return '2.8rem';
-      case 'xxl':
-        return '3.2rem';
-      default:
-        return size || '2.0rem';
-    }
-  }};
-  height: ${({ size }) => {
-    switch (size) {
-      case 'xs':
-        return '1.2rem';
-      case 'sm':
-        return '1.6rem';
-      case 'md':
-        return '2.0rem';
-      case 'lg':
-        return '2.4rem';
-      case 'xl':
-        return '2.8rem';
-      case 'xxl':
-        return '3.2rem';
-      default:
-        return size || '2.0rem';
-    }
-  }};
-  border: ${({ border }) => (border ? `${border}px` : '2px')} solid #e9ecef;
-  border-bottom-color: #212529bf;
-  border-radius: 50%;
-  display: inline-block;
-  box-sizing: border-box;
-  animation: rotation 1s linear infinite;
-
-  @keyframes rotation {
-    0% {
-      transform: rotate(0deg);
-    }
-    100% {
-      transform: rotate(360deg);
-    }
-  }
-
-  ${({ css }) => css}
-`;
+const Loader = ({ size, border, css }: LoaderProps) => {
+  return <StyledLoader size={size} border={border} css={css} />;
+};
 
 export default Loader;
