+++ frontend/src/components/common/Modal/Modal.tsx
@@ -1,7 +1,8 @@
 import type { CSSProp } from 'styled-components';
-import styled, { keyframes } from 'styled-components';
 
 import type { MouseEvent, ReactNode } from 'react';
+
+import { ModalContent, ModalWrapper } from './Modal.style';
 
 export interface ModalProps {
   isOpen: boolean;
@@ -58,43 +59,3 @@ const Modal = ({
 };
 
 export default Modal;
-
-const fadeIn = keyframes`
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-`;
-
-const ModalWrapper = styled.div<{ css: CSSProp }>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-
-  background: rgba(0, 0, 0, 0.5);
-  overflow: auto;
-
-  animation: ${fadeIn} 0.2s ease-in-out;
-
-  z-index: 9999;
-
-  ${({ css }) => css};
-`;
-
-const ModalContent = styled.div<{ noBackdrop: boolean; css: CSSProp }>`
-  background: #fff;
-  margin: 1rem;
-  border-radius: 10px;
-  max-height: calc(100% - 4rem);
-  overflow-y: auto;
-
-  ${({ noBackdrop, css }) => noBackdrop && css};
-`;
