+++ frontend/src/components/common/SelectBox/SelectBox.tsx
@@ -1,6 +1,6 @@
-import styled from 'styled-components';
-
 import { useState } from 'react';
+
+import { OptionItem, OptionsContainer, SelectButton, StyledSelect } from './SelectBox.style';
 
 interface SelectOption {
   value: string;
@@ -38,43 +38,5 @@ const SelectBox = ({ options, onChange, value }: SelectProps) => {
     </StyledSelect>
   );
 };
-
-const StyledSelect = styled.div`
-  position: relative;
-  width: 100%;
-`;
-
-const SelectButton = styled.button`
-  padding: 0.8rem;
-  font-size: 1.6rem;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  outline: none;
-  width: 100%;
-  background-color: #ffffff;
-  cursor: pointer;
-`;
-
-const OptionsContainer = styled.div`
-  position: absolute;
-  top: 100%;
-  left: 0;
-  width: 100%;
-  font-size: 1.6rem;
-  border: 1px solid #ccc;
-  border-top: none;
-  border-radius: 0 0 4px 4px;
-  background-color: #ffffff;
-  z-index: 1;
-`;
-
-const OptionItem = styled.div`
-  padding: 0.8rem;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #f0f0f0;
-  }
-`;
 
 export default SelectBox;
