+++ frontend/src/components/common/Text/Text.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta } from '@storybook/react';
-import { styled } from 'styled-components';
+import styled from 'styled-components';
 
 import Text from './Text';
