+++ frontend/src/components/ui/ToastContainer.tsx
@@ -2,8 +2,8 @@ import { useExternalValue } from '@utils/external-state';
 
 import { toastListStore } from '@stores/layout/toastStore';
 
-import type { ToastProps } from './Toast';
-import Toast from './Toast';
+import type { ToastProps } from '@common/Toast/Toast';
+import Toast from '@common/Toast/Toast';
 
 const ToastContainer = () => {
   const toastItems = useExternalValue<ToastProps[]>(toastListStore);
